+++ src/Divider/Divider.js
@@ -1,5 +1,5 @@
 import { Divider as ThemeUiDivider } from 'theme-ui';
 
 export default function Divider({ ...rest }) {
-  return <ThemeUiDivider {...rest} />;
+  return <ThemeUiDivider sx={{ color: 'divider' }} {...rest} />;
 }